refactor: removed SiteCountry, and substituted with SiteAddress which includes also other information about the address store
